Updated to support latest /dev/fuse DoInit() upcall
Also deprecated "darwin" support as it has fallen out-of-favor
  in terms of support from both {Mac|OSX|Fuse & Apple.

Also updated to a docker-compose dev environment.
