+++ api.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2015-2021, NVIDIA CORPORATION.
+// Copyright (c) 2015-2023, NVIDIA CORPORATION.
 // SPDX-License-Identifier: Apache-2.0
 
 package fission
@@ -93,8 +93,26 @@ func NewVolume(volumeName string, mountpointDirPath string, fuseSubtype string, 
 	return
 }
 
-// See api_{darwin|linux}.go for const AttrSize
-// See api_{darwin|linux}.go for type Attr struct
+const AttrSize = 88
+
+type Attr struct {
+	Ino       uint64
+	Size      uint64
+	Blocks    uint64
+	ATimeSec  uint64
+	MTimeSec  uint64
+	CTimeSec  uint64
+	ATimeNSec uint32
+	MTimeNSec uint32
+	CTimeNSec uint32
+	Mode      uint32
+	NLink     uint32
+	UID       uint32
+	GID       uint32
+	RDev      uint32
+	BlkSize   uint32
+	Padding   uint32
+}
 
 const KStatFSSize = 80
 
@@ -121,22 +139,18 @@ type FileLock struct {
 }
 
 const (
-	SetAttrInValidMode      = uint32(1) << 0
-	SetAttrInValidUID       = uint32(1) << 1
-	SetAttrInValidGID       = uint32(1) << 2
-	SetAttrInValidSize      = uint32(1) << 3
-	SetAttrInValidATime     = uint32(1) << 4
-	SetAttrInValidMTime     = uint32(1) << 5
-	SetAttrInValidFH        = uint32(1) << 6
-	SetAttrInValidATimeNow  = uint32(1) << 7
-	SetAttrInValidMTimeNow  = uint32(1) << 8
-	SetAttrInValidLockOwner = uint32(1) << 9
-	SetAttrInValidCTime     = uint32(1) << 10 // not supported
-
-	SetAttrInValidCrTime   = uint32(1) << 28 // darwin only
-	SetAttrInValidChgTime  = uint32(1) << 29 // darwin only
-	SetAttrInValidBkupTime = uint32(1) << 30 // darwin only
-	SetAttrInValidFlags    = uint32(1) << 31 // darwin only
+	SetAttrInValidMode        = uint32(1) << 0
+	SetAttrInValidUID         = uint32(1) << 1
+	SetAttrInValidGID         = uint32(1) << 2
+	SetAttrInValidSize        = uint32(1) << 3
+	SetAttrInValidATime       = uint32(1) << 4
+	SetAttrInValidMTime       = uint32(1) << 5
+	SetAttrInValidFH          = uint32(1) << 6
+	SetAttrInValidATimeNow    = uint32(1) << 7
+	SetAttrInValidMTimeNow    = uint32(1) << 8
+	SetAttrInValidLockOwner   = uint32(1) << 9
+	SetAttrInValidCTime       = uint32(1) << 10 // not supported
+	SetAttrInValidKillSuidGID = uint32(1) << 11 // not supported
 )
 
 const (
@@ -191,10 +205,15 @@ const (
 	InitFlagsNoOpendirSupport  = uint32(1) << 24
 	InitFlagsExplicitInvalData = uint32(1) << 25
 	InitFlagsMapAlignment      = uint32(1) << 26
+	InitFlagsSubMounts         = uint32(1) << 27
+	InitFlagsHandleKillprivV2  = uint32(1) << 28
+	InitFlagsSetXattrExt       = uint32(1) << 29
+	InitFalgsInitExt           = uint32(1) << 30
+	InitFlagsInitReserved      = uint32(1) << 31
 
-	InitFlagsCaseSensitive = uint32(1) << 29 // darwin only
-	InitFlagsVolRename     = uint32(1) << 30 // darwin only
-	InitFlagsXtimes        = uint32(1) << 31 // darwin only
+	InitFlags2SecurityCtx     = uint32(1) << 0
+	InitFlags2HasInodeDAX     = uint32(1) << 1
+	InitFlags2CreateSuppGroup = uint32(1) << 2
 )
 
 const (
@@ -277,7 +296,7 @@ type ForgetOne struct {
 
 const DirEntAlignment = 8 // applies to both DirEnt and DirEntPlus
 
-const DirEntFixedPortionSize = 24 // + len(Name) and rounded up to DirEntAlignment boundary
+const DirEntFixedPortionSize = 24 // + le251,255,255,115n(Name) and rounded up to DirEntAlignment boundary
 
 type DirEnt struct {
 	Ino     uint64
@@ -344,6 +363,8 @@ const (
 	OpCodeCopyFileRange = uint32(47) // unsupported
 	OpCodeSetupMapping  = uint32(48) // unsupported
 	OpCodeRemoveMapping = uint32(49) // unsupported
+	OpCodeSyncFS        = uint32(50) // unsupported
+	OpCodeTmpFile       = uint32(51) // unsupported
 
 	OpCodeCuseInit = uint32(4096) // unsupported
 )
@@ -402,8 +423,26 @@ type GetAttrOut struct {
 	Attr
 }
 
-// See api_{darwin|linux}.go for const SetAttrInSize
-// See api_{darwin|linux}.go for type SetAttrIn struct
+const SetAttrInSize = 88
+
+type SetAttrIn struct {
+	Valid     uint32 // mask of const SetAttrInValid* bits
+	Padding   uint32
+	FH        uint64
+	Size      uint64
+	LockOwner uint64
+	ATimeSec  uint64
+	MTimeSec  uint64
+	Unused2   uint64
+	ATimeNSec uint32
+	MTimeNSec uint32
+	Unused3   uint32
+	Mode      uint32
+	Unused4   uint32
+	UID       uint32
+	GID       uint32
+	Unused5   uint32
+}
 
 const SetAttrOutSize = 16 + AttrSize
 
@@ -562,25 +601,23 @@ type FSyncIn struct {
 	Padding    uint32
 }
 
-// See api_{darwin|linux}.go for const SetXAttrInFixedPortionSize
+const SetXAttrInFixedPortionSize = 16 // + len(Name) + 1 + len(Data)
 
 type SetXAttrIn struct {
-	Size     uint32 // == len(Name) + 1 + len(Data)
-	Flags    uint32 // if not 0, either of SetXAttrIn*
-	Position uint32 // darwin only - set to zero otherwise
-	Padding  uint32 // darwin only
-	Name     []byte
-	Data     []byte // byte(0) separated from Name
+	Size          uint32 // == len(Name) + 1 + len(Data)
+	Flags         uint32 // if not 0, either of SetXAttrIn*
+	SetXAttrFlags uint32
+	Padding       uint32
+	Name          []byte
+	Data          []byte // byte(0) separated from Name
 }
 
-// See api_{darwin|linux}.go for const GetXAttrInFixedPortionSize
+const GetXAttrInFixedPortionSize = 8 // + len(Name)
 
 type GetXAttrIn struct {
-	Size     uint32 // == max len(GetXAttrOut.Data)
-	Padding  uint32
-	Position uint32 // darwin only - set to zero otherwise
-	Padding2 uint32 // darwin only
-	Name     []byte
+	Size    uint32 // == max len(GetXAttrOut.Data)
+	Padding uint32
+	Name    []byte
 }
 
 const GetXAttrOutSizeOnlySize = 8
@@ -591,12 +628,11 @@ type GetXAttrOut struct {
 	Data    []byte // only returned if GetXAttrIn.Size != 0
 }
 
-// See api_{darwin|linux}.go for const ListXAttrInSize
+const ListXAttrInSize = 8
 
 type ListXAttrIn struct {
-	Size     uint32 // == max len(ListXAttrOut.Name) with a '\0' between each Name element
-	Padding  uint32
-	Padding2 uint64 // darwin only
+	Size    uint32 // == max len(ListXAttrOut.Name) with a '\0' between each Name element
+	Padding uint32
 }
 
 const ListXAttrOutSizeOnlySize = 8
@@ -620,16 +656,17 @@ type FlushIn struct {
 	LockOwner uint64
 }
 
-const InitInSize = 16
+const InitInSize = 64
 
 type InitIn struct {
 	Major        uint32
 	Minor        uint32
 	MaxReadAhead uint32
 	Flags        uint32 // mask of const InitFlags* bits
+	Flags2       uint32 // mask of const InitFlags2* bits
+	Unused       [11]uint32
 }
 
-const InitOutSizePre_7_28 = 24
 const InitOutSize = 64
 
 type InitOut struct {
@@ -640,10 +677,11 @@ type InitOut struct {
 	MaxBackground        uint16
 	CongestionThreshhold uint16
 	MaxWrite             uint32
-	TimeGran             uint32    // starting at 7.28
-	MaxPages             uint16    // starting at 7.28
-	Padding              uint16    // starting at 7.28
-	Unused               [8]uint32 // starting at 7.28
+	TimeGran             uint32
+	MaxPages             uint16
+	MapAlignment         uint16
+	Flags2               uint32
+	Unused               [7]uint32
 }
 
 const OpenDirInSize = 8
