--- examples/fission-swiftfs/attr_darwin.go
@@ -1,17 +1,0 @@
-// Copyright (c) 2015-2021, NVIDIA CORPORATION.
-// SPDX-License-Identifier: Apache-2.0
-
-package main
-
-import (
-	"github.com/NVIDIA/fission"
-)
-
-func fixAttr(attr *fission.Attr) {
-	attr.Blocks = (attr.Size + uint64(attrBlkSize) - 1) / uint64(attrBlkSize)
-	attr.BlkSize = attrBlkSize
-
-	attr.CrTimeSec = attr.CTimeSec
-	attr.CrTimeNSec = attr.CTimeNSec
-	attr.Flags = 0
-}
